refactor: better package organisation
Created `pkg` and moved all root code there. Also internal was discarded
and code was moved into `pkg/tui`.

I've decided to change packages because kconf is primarly CLI
application, not a library. Using kconf as library is possible but is
very rare use-case.
