+++ cmd/kconf/rm/rm.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/sn3d/kconf"
+	"github.com/sn3d/kconf/pkg/kconf"
 	"github.com/urfave/cli/v2"
 )
