+++ cmd/kconf/split/split.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/sn3d/kconf"
+	"github.com/sn3d/kconf/pkg/kconf"
 	"github.com/urfave/cli/v2"
 )
