+++ pkg/kconf/testdata/remove-test.yaml
