+++ pkg/kconf/testdata/rename-test.yaml
